mqtt_worker/requirements.txt
paho-mqtt==1.6.1
meshtastic==2.3.0
protobuf==4.25.1
h3==3.7.6
